backend/requirements.txt
Django==4.2.20
djangorestframework==3.16.0
psycopg2-binary==2.9.9
openai==1.35.3
python-docx==1.1.2
python-dotenv==1.0.1
httpx==0.27.2